app/requirements.txt
fastapi
uvicorn
langchain-experimental
open_clip_torch
pillow
pydantic
python-multipart